Moving all utility functions to the utils module
